rest/requirements.txt
requests==2.27.1
Vega-API-client==0.49.2
websocket-client==1.3.2
